Revert "patch(hardhat-core,hardhat-ethers): moved linkBytecode function & library related typing to hardhat-core"
This reverts commit 18740ec67a446bc783eb96001cdd82c33db33cea.
