+++ packages/hardhat-core/src/types/index.ts
@@ -4,4 +4,3 @@ export * from "./provider";
 export * from "./runtime";
 export * from "./artifacts";
 export * from "./builtin-tasks";
-export * from "./libraries";
