--- packages/hardhat-core/src/types/libraries.ts
@@ -1,9 +1,0 @@
-export interface Libraries<Address = string> {
-  [libraryName: string]: Address;
-}
-
-export interface Link {
-  sourceName: string;
-  libraryName: string;
-  address: string;
-}
