+++ packages/hardhat-viem/src/internal/bytecode.ts
@@ -1,8 +1,6 @@
 import type * as viemT from "viem";
 
 import { Artifact } from "hardhat/types/artifacts";
-import { Libraries, Link } from "hardhat/types/libraries";
-import { linkBytecode } from "hardhat/utils/link-bytecode";
 
 import {
   AmbigousLibraryNameError,
@@ -11,6 +9,33 @@ import {
   OverlappingLibraryNamesError,
   UnnecessaryLibraryLinkError,
 } from "./errors";
+
+type Libraries<Address = string> = {
+  [libraryName: string]: Address;
+};
+
+interface Link {
+  sourceName: string;
+  libraryName: string;
+  address: string;
+}
+
+export function linkBytecode(artifact: Artifact, libraries: Link[]): string {
+  let bytecode = artifact.bytecode;
+
+  // TODO: measure performance impact
+  for (const { sourceName, libraryName, address } of libraries) {
+    const linkReferences = artifact.linkReferences[sourceName][libraryName];
+    for (const { start, length } of linkReferences) {
+      bytecode =
+        bytecode.substring(0, 2 + start * 2) +
+        address.substring(2) +
+        bytecode.substring(2 + (start + length) * 2);
+    }
+  }
+
+  return bytecode;
+}
 
 export async function throwOnAmbigousLibraryNameOrUnnecessaryLink(
   contractName: string,
